ROC-1183 - Separated DynamicMessage functions from dialectRT.go into dynamic_message.go.
